New DiscreteBilinearForm not subclass of BasicCodeGen
